segurizacion de enpoint

+++ microservicio_viaje/src/main/java/com/microservicio_viaje/microservicio_viaje/controller/TarifaController.java
@@ -2,7 +2,9 @@ package com.microservicio_viaje.microservicio_viaje.controller;
 
 import com.microservicio_viaje.microservicio_viaje.model.Tarifa;
 import com.microservicio_viaje.microservicio_viaje.repository.TarifaRepository;
+import com.microservicio_viaje.microservicio_viaje.service.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,14 +15,22 @@ import java.util.List;
 public class TarifaController {
     @Autowired
     private TarifaRepository tarifaRepository;
+    @Autowired
+    JwtService jwtService;
 
     /**
      *
      * @return Muestro todas las tarifas
      */
     @GetMapping("/tarifa")
-    public List<Tarifa> getAll(){
-        return tarifaRepository.findAll();
+    public List<Tarifa> getAll(@RequestHeader("Authorization") String authorizationHeader){
+        String token = authorizationHeader.replace("Bearer ", "");
+        if(jwtService.isTokenValid(token)) {
+            return tarifaRepository.findAll();
+        }else{
+            ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+            return null;
+        }
     }
 
     /**
@@ -29,12 +39,14 @@ public class TarifaController {
      * @return
      */
     @PostMapping("/CreoUnaTarifa")
-    public ResponseEntity<String> create(@RequestBody Tarifa tarifa){
-        try {
+    public ResponseEntity<String> create(@RequestHeader("Authorization") String authorizationHeader, @RequestBody Tarifa tarifa){
+        String token = authorizationHeader.replace("Bearer ", "");
+        if(jwtService.isTokenValid(token)) {
             tarifaRepository.save(tarifa);
             return ResponseEntity.ok("Se agregó con exito");
-        }catch (Exception e){
-            return ResponseEntity.badRequest().body("Lamentablemente no se posible agregar el dato "+e.getMessage());
+        }else{
+            ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+            return null;
         }
     }
 
@@ -43,8 +55,14 @@ public class TarifaController {
      * @return tarifa asociada a la id
      */
     @GetMapping("/tarifa/id/{id}")
-    public Tarifa get(@PathVariable int id){
-        return tarifaRepository.findById(id).orElse(null);
+    public Tarifa get(@RequestHeader("Authorization") String authorizationHeader, @PathVariable int id){
+        String token = authorizationHeader.replace("Bearer ", "");
+        if(jwtService.isTokenValid(token)) {
+            return tarifaRepository.findById(id).orElse(null);
+        }else{
+            ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+            return null;
+        }
     }
 
     /**
@@ -53,17 +71,25 @@ public class TarifaController {
      * @return una repuesta del servidor
      */
     @DeleteMapping("eliminarTarifa/{id}")
-    public ResponseEntity<String> delete(@PathVariable int id){
-        try{
+    public ResponseEntity<String> delete(@RequestHeader("Authorization") String authorizationHeader, @PathVariable int id){
+        String token = authorizationHeader.replace("Bearer ", "");
+        if(jwtService.isTokenValid(token)) {
             tarifaRepository.deleteById(id);
             return ResponseEntity.ok("Se elimino con exito");
-        }catch (Exception e){
-            return ResponseEntity.badRequest().body("No se pudo eliminar "+e.getMessage());
+        }else{
+            ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+            return null;
         }
     }
     @GetMapping("/tipo/{tipo}")
-    public double getPrecioPorMinuto(@PathVariable String tipo) {
-        return tarifaRepository.findTarifaByTipo(tipo);
+    public double getPrecioPorMinuto(@RequestHeader("Authorization") String authorizationHeader, @PathVariable String tipo) {
+        String token = authorizationHeader.replace("Bearer ", "");
+        if(jwtService.isTokenValid(token)) {
+            return tarifaRepository.findTarifaByTipo(tipo);
+        }else{
+            ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+            return 0.0;
+        }
     }
 
     /**
@@ -72,16 +98,16 @@ public class TarifaController {
      * @param valor
      */
     @PutMapping("actualizarTarifa/{id}/{valor}")
-    public ResponseEntity<String> actualizarTarifa(@PathVariable int id, @PathVariable int valor){
+    public ResponseEntity<String> actualizarTarifa(@RequestHeader("Authorization") String authorizationHeader, @PathVariable int id, @PathVariable int valor){
+        String token = authorizationHeader.replace("Bearer ", "");
+        if(jwtService.isTokenValid(token)) {
         Tarifa tarifa = tarifaRepository.findById(id).orElse(null);
-        try{
             tarifa.setValor(valor);
             tarifaRepository.save(tarifa);
             return ResponseEntity.ok("El valor de la tarifa "+tarifa.getTipo()+" se actualizó correctamente");
-        }catch (Exception e){
-            return ResponseEntity.badRequest().body("La tarifa que está queriendo actualizar no exite.");
+        }else{
+            ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+            return null;
         }
     }
-    
-
 }
